convert functional component to the class component
